refactor : 바 이미지 등록 컴포넌트 분리, 리렌더 간섭 방지
